Issues with mass, tselecopts, and time
Mass was being passed to rungtd1d but mass wasn't an input for rungtd1d.  Was running into a problem where tselecopts wasn't defined so it wouldn't work. Not sure what that does though. For the altitude 1-D case had to select the first element of dtime or assert isinstance(t,(datetime,str)) wouldn't work